refactor(widget): share one polygon-anchor helper across map views
